Moved and improved AxisResizer.
